remove parallelfor.cpp
